Migrate first quantized prepare.
