+++ qualtran/bloqs/chemistry/pbc/first_quantization/prepare.py
@@ -1,0 +1,57 @@
+#  Copyright 2023 Google LLC
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+r"""PREPARE for the first quantized chemistry Hamiltonian."""
+from functools import cached_property
+from typing import Optional, Set, Tuple, TYPE_CHECKING
+
+from attrs import frozen
+
+from qualtran import Bloq, Signature
+from qualtran.bloqs.basic_gates import Toffoli
+
+if TYPE_CHECKING:
+    from qualtran.resource_counting import SympySymbolAllocator
+
+
+@frozen
+class UniformSuperPostionIJFirstQuantization(Bloq):
+    r"""Uniform superposition over $\eta$ values of $i$ and $j$ in unary such that $i \ne j$.
+
+    Args:
+        eta: The number of electrons.
+        num_bits_rot_aa: The number of bits of precision for the single qubit
+            rotation for amplitude amplification. Called $b_r$ in the reference.
+        adjoint: whether to dagger the bloq or not.
+
+    Registers:
+        i: a n_eta bit register for unary encoding of eta numbers.
+        j: a n_eta bit register for unary encoding of eta numbers.
+
+    References:
+        [Fault-Tolerant Quantum Simulations of Chemistry in First Quantization]
+        (https://arxiv.org/abs/2105.12767) page 18, section A, around Eq 62.
+    """
+    eta: int
+    num_bits_rot_aa: int
+    adjoint: int = False
+
+    @cached_property
+    def signature(self) -> Signature:
+        n_eta = (self.eta - 1).bit_length()
+        return Signature.build(i=n_eta, j=n_eta)
+
+    def bloq_counts(self, ssa: Optional['SympySymbolAllocator'] = None) -> Set[Tuple[int, Bloq]]:
+        n_eta = (self.eta - 1).bit_length()
+        # Half of Eq. 62 which is the cost for prep and prep^\dagger
+        return {((7 * n_eta + 4 * self.num_bits_rot_aa - 18), Toffoli())}
