+++ qualtran/bloqs/chemistry/pbc/first_quantization/prepare_nu.py
@@ -11,73 +11,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
-r"""SELECT and PREPARE for the first quantized chemistry Hamiltonian.
-
-Here we assume the Born-Oppenheimer Hamiltonian and seek to simulate a
-collection of $\eta$ electrons and $L$ static nuclei with a Hamiltonian given
-by:
-$$
-H_{BO} = T + U + V + \frac{1}{2}
-\sum_{\ell\ne\kappa=1}^L\frac{\zeta_\ell\zeta_\kappa}{\lVert R_\ell - R_\kappa \rVert}
-$$
-
-In the first quantized approach we assume periodic boundary conditions and use a
-plane wave Galerkin discretization.
-A plane wave basis function is given by
-$$
-\phi_p(r) = \frac{1}{\sqrt{\Omega}} e^{-i k_p\cdot r}
-$$
-where $r$ is a position vector in real space, $\Omega$ is the simulation cell
-volume and $k_p$ is a reciprocal lattice vector.
-In three dimensions we have
-$$
-k_p = \frac{2\pi p }{\Omega}
-$$
-for $p \in G$ and
-$$
-G = [-\frac{N^{1/3} -
-1}{2},{N^{1/3} - 1}{2}]^3 \subset \mathcal{Z}^3.
-$$
-and $N$ is the total number of planewaves.
-
-With these definitions we can write the components of the Hamiltonian as:
-$$
-T = \sum_{i}^\eta\sum_{p\in G}\frac{\lVert k_p\rVert^2}{2} |p\rangle\langle p|_i
-$$
-which defines the kinetic energy of the electrons,
-$$
-U = -\frac{4\pi}{\Omega}
-\sum_{\ell=1}^L \sum_{i}^\eta
-\sum_{p,q\in G, p\ne q}
-\left(
-    \zeta_{\ell}
-    \frac{e^{i k_{q-p}\cdot R_\ell}}{\lVert k_{p-q}\rVert^2}
-    |p\rangle\langle q|_i
-\right)
-$$
-describes the interaction of the electrons and the nuclei, and,
-$$
-V = \frac{2\pi}{\Omega}
-\sum_{i\nej=1}^\eta
-\sum_{p,q\in G, p\ne q}
-\sum_{\nu \in G_0}
-\left(
-    \frac{1}{\lVert k_{\nu}\rVert^2}
-    |p + \nu\rangle\langle p|_i
-    |q -\nu\rangle\langle q|_i
-\right)
-$$
-describes the electron-electron interaction. The notation $|p\rangle\langle p|i$ is shorthand for
-$I_1\otensor\cdots\otensor |p\rangle \langle q |_j \otensor \cdots \otensor I_\eta$.
-The system is represented using a set of $\eta$ signed integer registers each of
-size $3 n_p$ where $n_p =  \lceil \log (N^{1/3} + 1) \rceil$, with the factor of
-3 accounting for the 3 spatial dimensions.
-
-In the first quantized approach, fermion antisymmetry is encoded through initial
-state preparation. Spin labels are also absent and should be accounted for
-during state preparation. The cost of initial state preparation is typically
-ignored.
-"""
+r"""Bloqs for preparation of the U and V parts of the first quantized chemistry Hamiltonian."""
 from functools import cached_property
 from typing import Optional, Set, Tuple, TYPE_CHECKING
 
@@ -90,123 +24,6 @@ from qualtran.bloqs.basic_gates import Toffoli
 
 if TYPE_CHECKING:
     from qualtran.resource_counting import SympySymbolAllocator
-
-
-@frozen
-class UniformSuperPostionIJFirstQuantization(Bloq):
-    r"""Uniform superposition over $\eta$ values of $i$ and $j$ in unary such that $i \ne j$.
-
-    Args:
-        eta: The number of electrons.
-        num_bits_rot_aa: The number of bits of precision for the single qubit
-            rotation for amplitude amplification. Called $b_r$ in the reference.
-        adjoint: whether to dagger the bloq or not.
-
-    Registers:
-        i: a n_eta bit register for unary encoding of eta numbers.
-        j: a n_eta bit register for unary encoding of eta numbers.
-
-    References:
-        [Fault-Tolerant Quantum Simulations of Chemistry in First Quantization]
-        (https://arxiv.org/abs/2105.12767) page 18, section A, around Eq 62.
-    """
-    eta: int
-    num_bits_rot_aa: int
-    adjoint: int = False
-
-    @cached_property
-    def signature(self) -> Signature:
-        n_eta = (self.eta - 1).bit_length()
-        return Signature.build(i=n_eta, j=n_eta)
-
-    def bloq_counts(self, ssa: Optional['SympySymbolAllocator'] = None) -> Set[Tuple[int, Bloq]]:
-        n_eta = (self.eta - 1).bit_length()
-        # Half of Eq. 62 which is the cost for prep and prep^\dagger
-        return {((7 * n_eta + 4 * self.num_bits_rot_aa - 18), Toffoli())}
-
-
-@frozen
-class PreparePowerTwoState(Bloq):
-    r"""Prepares the uniform superposition over $|r\rangle$ given by Eq. 69 in the reference.
-
-    This prepares the state
-
-    $$
-        2^{(-n_p -1)/2} \sum_r=0^{n_p-2} 2^{r/2} |r\rangle
-    $$
-
-    in one-hot unary.
-
-    Args:
-        bitsize: the number of bits $n_p$ for the $|r\rangle$ register.
-
-    Registers:
-        r: The register we want to prepare the state over.
-
-    References:
-        [Fault-Tolerant Quantum Simulations of Chemistry in First Quantization]
-        (https://arxiv.org/abs/2105.12767) Eq 67-69, pg 19-20
-    """
-    bitsize: int
-
-    @cached_property
-    def signature(self) -> Signature:
-        return Signature.build(r=self.bitsize)
-
-    def bloq_counts(self, ssa: Optional['SympySymbolAllocator'] = None) -> Set[Tuple[int, Bloq]]:
-        return {((self.bitsize - 2), Toffoli())}
-
-
-@frozen
-class PrepareTFirstQuantization(Bloq):
-    r"""PREPARE for the kinetic energy operator for the first quantized chemistry Hamiltonian.
-
-    This prepares the state
-
-    $$
-        |+\rangle\sum_{j=1}^{\eta}|j\rangle\sum_{w=0}^{2}|w\rangle
-        \sum_{r=0}^{n_{p}-2}2^{r/2}|r\rangle
-        \sum_{s=0}^{n_{p}-2}2^{s/2}|s\rangle
-    $$
-
-    We assume that the uniform superposition over ($i$ and) $j$ has already occured via
-    UniformSuperPositionIJFirstQuantization.
-
-    Args:
-        num_bits_p: The number of bits to represent each dimension of the momentum register.
-        eta: The number of electrons.
-        num_bits_rot_aa: The number of bits of precision for the single qubit
-            rotation for amplitude amplification. Called $b_r$ in the reference.
-        adjoint: whether to dagger the bloq or not.
-
-    Registers:
-        plus: A $|+\rangle$ state register.
-        w: a register to index one of three components of the momenta.
-        r: a register encoding bits for each component of the momenta.
-        s: a register encoding bits for each component of the momenta.
-
-    References:
-        [Fault-Tolerant Quantum Simulations of Chemistry in First Quantization](
-            https://arxiv.org/abs/2105.12767) page 19, section B
-    """
-
-    num_bits_p: int
-    eta: int
-    num_bits_rot_aa: int = 8
-    adjoint: int = False
-
-    @cached_property
-    def signature(self) -> Signature:
-        return Signature.build(plus=1, w=2, r=self.num_bits_p - 2, s=self.num_bits_p - 2)
-
-    def bloq_counts(self, ssa: Optional['SympySymbolAllocator'] = None) -> Set[Tuple[int, Bloq]]:
-        # there is a cost for the uniform state preparation for the $w$
-        # register. Adding a bloq is sort of overkill, should just tag the
-        # correct cost on UniformSuperPosition bloq
-        # 13 is from assuming 8 bits for the rotation, and n = 2.
-        uni_prep_w = (13, Toffoli())
-        # Factor of two for r and s registers.
-        return {uni_prep_w, (2, PreparePowerTwoState(bitsize=self.num_bits_p))}
 
 
 @frozen
@@ -471,104 +288,3 @@ class PrepareNuState(Bloq):
         # 6. Test that $(2^{\mu-2})^2\mathcal{M} > m (\nu_x^2 + \nu_y^2 + \nu_z^2)$
         cost_6 = (1, TestNuInequality(self.num_bits_p, n_m, adjoint=self.adjoint))
         return {cost_1, cost_2, cost_3, cost_4, cost_6}
-
-
-@frozen
-class PrepareZetaState(Bloq):
-    r"""PREPARE the superpostion over $l$ weighted by $\zeta_l$.
-
-    This is apparently NOT just generic state preparation and there are some
-    tricks I don't understand.
-
-    Args:
-        num_bits_p: The number of bits to represent each dimension of the momentum register.
-        eta: The number of electrons.
-        m_param: $\mathcal{M}$ in the reference.
-        lambda_zeta: sum of nuclear charges.
-        er_lambda_zeta: eq 91 of the referce. Cost of erasing qrom.
-
-    Registers:
-
-    References:
-        [Fault-Tolerant Quantum Simulations of Chemistry in First Quantization](
-            https://arxiv.org/abs/2105.12767) page 23-24, last 3 paragraphs.
-    """
-    num_atoms: int
-    lambda_zeta: int
-    num_bits_nuc_pos: int
-    adjoint: bool = False
-
-    @cached_property
-    def signature(self) -> Signature:
-        return Signature([Register("l", bitsize=(self.num_atoms - 1).bitsize())])
-
-    def bloq_counts(self, ssa: Optional['SympySymbolAllocator'] = None) -> Set[Tuple[int, Bloq]]:
-        if self.adjoint:
-            # Really Er(x), eq 91. In practice we will reaplce this with the
-            # appropriate qrom call down the line.
-            return {(int(np.ceil(self.lambda_zeta**0.5)), Toffoli())}
-        else:
-            return {(self.lambda_zeta, Toffoli())}
-
-
-@frozen
-class PrepareUVFistQuantization(Bloq):
-    r"""PREPARE the U and V parts of the Hamiltonian.
-
-    Args:
-        num_bits_p: The number of bits to represent each dimension of the momentum register.
-        eta: The number of electrons.
-        num_atoms: The number of atoms. $L$ in the reference.
-        m_param: $\mathcal{M}$ in the reference.
-        lambda_zeta: sum of nuclear charges.
-        num_bits_nuc_pos: The number of bits of precision for representing the nuclear coordinates.
-
-    Registers:
-        mu: The state controlling the nested boxes procedure.
-        nu: The momentum transfer register.
-        m: an ancilla register in a uniform superposition.
-        l: The register for atomic species.
-        flag_nu: Flag for success of the state preparation.
-
-    References:
-        [Fault-Tolerant Quantum Simulations of Chemistry in First Quantization](
-            https://arxiv.org/abs/2105.12767) page 19, section B
-    """
-    num_bits_p: int
-    eta: int
-    num_atoms: int
-    m_param: int
-    lambda_zeta: int
-    num_bits_nuc_pos: int
-    adjoint: bool = False
-
-    @cached_property
-    def signature(self) -> Signature:
-        # this is for the nu register which lives on a grid of twice the size
-        # the nu grid is twice as large, so one more bit is needed
-        n_nu = (self.num_bits_p - 1).bit_length() + 2
-        n_mu = self.num_bits_p.bit_length()
-        #
-        n_m = (self.m_param - 1).bit_length()
-        return Signature(
-            [
-                Register("mu", bitsize=n_mu),
-                Register("nu", bitsize=n_nu, shape=(3,)),
-                Register("m", bitsize=n_m),
-                Register("l", bitsize=self.num_bits_nuc_pos),
-                Register("flag_nu", bitsize=1),
-            ]
-        )
-
-    def bloq_counts(self, ssa: Optional['SympySymbolAllocator'] = None) -> Set[Tuple[int, Bloq]]:
-        # 1. Prepare the nu state
-        # 2. Prepare the zeta_l state
-        return {
-            (1, PrepareNuState(self.num_bits_p, self.m_param, self.adjoint)),
-            (
-                1,
-                PrepareZetaState(
-                    self.num_atoms, self.lambda_zeta, self.num_bits_nuc_pos, self.adjoint
-                ),
-            ),
-        }
