+++ qualtran/bloqs/chemistry/pbc/first_quantization/prepare_test.py
@@ -1,0 +1,20 @@
+#  Copyright 2023 Google LLC
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+from qualtran.bloqs.chemistry.pbc.first_quantizationzation.prepare import (
+    UniformSuperPostionIJFirstQuantization,
+)
+
+
+def test_uniform_super_position_ij():
+    prep = UniformSuperPostionIJFirstQuantization(eta=10, num_bits_rot_aa=8)
